layers: Move Android CoreChecks to own file
