+++ BUILD.gn
@@ -109,6 +109,7 @@ config("vulkan_layer_config") {
 
 core_validation_sources = [
   "layers/android_ndk_types.h",
+  "layers/android_validation.cpp",
   "layers/base_node.cpp",
   "layers/base_node.h",
   "layers/buffer_state.cpp",
